+++ mqtt/Cargo.toml
@@ -7,6 +7,7 @@ members = [
     "mqtt-edgehub",
     "mqtt-generic",
     "mqtt-policy",
+    "mqtt-util",
     "mqttd",
     "policy"
 ]
