+++ mqtt/mqtt-bridge/Cargo.toml
@@ -11,15 +11,12 @@ bincode = "1.2"
 bson = "1.1"
 bytes = "0.5"
 config = { version = "0.10", features = ["json"], default-features = false }
-chrono = "0.4"
 futures-util = "0.3"
 humantime = "2.0"
 humantime-serde = "1.0"
 lazy_static = "1.4"
 mockall = "0.8"
-openssl = "0.10"
 parking_lot = "0.11"
-percent-encoding = "1.0"
 regex = "1"
 serde = { version = "1.0", features = ["derive", "rc"] }
 serde_bytes = "0.11"
@@ -28,8 +25,6 @@ serde_json = "1.0"
 serial_test = "0.4"
 thiserror = "1.0"
 tokio = { version = "0.2", features = ["sync", "rt-util"] }
-tokio-tls = "0.3"
-tokio-openssl = "0.4"
 tracing = "0.1"
 tracing-futures = "0.2"
 url = "2.1"
@@ -37,6 +32,7 @@ url = "2.1"
 edgelet-client = { path = "../edgelet-client" }
 mqtt3 = { path = "../mqtt3", features = ["serde1"] }
 mqtt-broker = { path = "../mqtt-broker" }
+mqtt-util = { path = "../mqtt-util" }
 
 [dev-dependencies]
 matches = "0.1"
