+++ mqtt/mqtt-bridge/src/client.rs
@@ -1,190 +1,26 @@
 #![allow(clippy::default_trait_access)] // Needed because mock! macro violates
 use core::{convert::TryInto, num::TryFromIntError};
-use std::{fmt::Display, io::Error, io::ErrorKind, pin::Pin, str, time::Duration};
+use std::{fmt::Display, str, time::Duration};
 
 use async_trait::async_trait;
-use chrono::Utc;
-use futures_util::future::{self, BoxFuture};
 use mockall::automock;
-use openssl::{ssl::SslConnector, ssl::SslMethod, x509::X509};
-use tokio::{io::AsyncRead, io::AsyncWrite, net::TcpStream, stream::StreamExt};
+use tokio::stream::StreamExt;
 use tracing::{debug, error, info};
 
 use mqtt3::{
     proto::{self, Publication, SubscribeTo},
-    Client, Event, IoSource, ShutdownError, UpdateSubscriptionError,
+    Client, Event, ShutdownError, UpdateSubscriptionError,
 };
 
-use crate::{
-    settings::Credentials,
-    token_source::{SasTokenSource, TokenSource, TrustBundleSource},
+use mqtt_util::client_io::{
+    ClientIoSource, Credentials, SasTokenSource, TcpConnection, TrustBundleSource,
 };
 
-const DEFAULT_TOKEN_DURATION_MINS: i64 = 60;
 const DEFAULT_MAX_RECONNECT: Duration = Duration::from_secs(60);
 // TODO: get QOS from topic settings
 const DEFAULT_QOS: proto::QoS = proto::QoS::AtLeastOnce;
 // TODO: read from env var
 const API_VERSION: &str = "2010-01-01";
-
-#[derive(Clone)]
-enum BridgeIoSource {
-    Tcp(TcpConnection<SasTokenSource>),
-    Tls(TcpConnection<SasTokenSource>),
-}
-
-trait BridgeIo: AsyncRead + AsyncWrite + Send + Sync + 'static {}
-
-impl<I> BridgeIo for I where I: AsyncRead + AsyncWrite + Send + Sync + 'static {}
-
-type BridgeIoSourceFuture =
-    BoxFuture<'static, Result<(Pin<Box<dyn BridgeIo>>, Option<String>), Error>>;
-
-#[derive(Clone)]
-pub struct TcpConnection<T>
-where
-    T: TokenSource + Clone + Send + Sync + 'static,
-{
-    address: String,
-    token_source: Option<T>,
-    trust_bundle_source: Option<TrustBundleSource>,
-}
-
-impl<T> TcpConnection<T>
-where
-    T: TokenSource + Clone + Send + Sync + 'static,
-{
-    pub fn new(
-        address: impl Into<String>,
-        token_source: Option<T>,
-        trust_bundle_source: Option<TrustBundleSource>,
-    ) -> Self {
-        Self {
-            address: address.into(),
-            token_source,
-            trust_bundle_source,
-        }
-    }
-}
-
-impl IoSource for BridgeIoSource {
-    type Io = Pin<Box<dyn BridgeIo>>;
-
-    type Error = Error;
-
-    #[allow(clippy::type_complexity)]
-    type Future = BoxFuture<'static, Result<(Self::Io, Option<String>), Self::Error>>;
-
-    fn connect(&mut self) -> Self::Future {
-        match self {
-            BridgeIoSource::Tcp(connect_settings) => Self::get_tcp_source(connect_settings.clone()),
-            BridgeIoSource::Tls(connect_settings) => Self::get_tls_source(connect_settings.clone()),
-        }
-    }
-}
-
-impl BridgeIoSource {
-    fn get_tcp_source(connection_settings: TcpConnection<SasTokenSource>) -> BridgeIoSourceFuture {
-        let address = connection_settings.address;
-        let token_source = connection_settings.token_source;
-
-        Box::pin(async move {
-            let expiry = Utc::now() + chrono::Duration::minutes(DEFAULT_TOKEN_DURATION_MINS);
-
-            let io = TcpStream::connect(&address);
-
-            let token_task = async {
-                match token_source {
-                    Some(ts) => ts.get(&expiry).await,
-                    None => Ok(None),
-                }
-            };
-
-            let (password, io) = future::try_join(token_task, io).await.map_err(|err| {
-                Error::new(ErrorKind::Other, format!("failed to connect: {}", err))
-            })?;
-
-            if let Some(pass) = &password {
-                validate_length(pass).map_err(|e| {
-                    error!(error = %e, "password too long");
-                    ErrorKind::InvalidInput
-                })?;
-            }
-
-            let stream: Pin<Box<dyn BridgeIo>> = Box::pin(io);
-            Ok((stream, password))
-        })
-    }
-
-    fn get_tls_source(connection_settings: TcpConnection<SasTokenSource>) -> BridgeIoSourceFuture {
-        let address = connection_settings.address.clone();
-        let token_source = connection_settings.token_source.as_ref().cloned();
-        let trust_bundle_source = connection_settings.trust_bundle_source;
-
-        Box::pin(async move {
-            let expiry = Utc::now() + chrono::Duration::minutes(DEFAULT_TOKEN_DURATION_MINS);
-
-            let server_root_certificate_task = async {
-                match trust_bundle_source {
-                    Some(source) => source.get_trust_bundle().await,
-                    None => Ok(None),
-                }
-            };
-
-            let token_task = async {
-                match token_source {
-                    Some(ts) => ts.get(&expiry).await,
-                    None => Ok(None),
-                }
-            };
-
-            let io = TcpStream::connect(address.clone());
-
-            let (server_root_certificate, password, stream) =
-                future::try_join3(server_root_certificate_task, token_task, io)
-                    .await
-                    .map_err(|err| {
-                        Error::new(ErrorKind::Other, format!("failed to connect: {}", err))
-                    })?;
-
-            if let Some(pass) = password.as_ref() {
-                validate_length(pass).map_err(|e| {
-                    error!(error = %e, "password too long");
-                    ErrorKind::InvalidInput
-                })?;
-            }
-
-            let config = SslConnector::builder(SslMethod::tls())
-                .map(|mut builder| {
-                    if let Some(trust_bundle) = server_root_certificate {
-                        X509::stack_from_pem(trust_bundle.as_bytes())
-                            .map(|mut certs| {
-                                while let Some(ca) = certs.pop() {
-                                    builder.cert_store_mut().add_cert(ca).ok();
-                                }
-                            })
-                            .ok();
-                    }
-
-                    builder.build()
-                })
-                .and_then(|conn| conn.configure())
-                .map_err(|e| Error::new(ErrorKind::NotConnected, e))?;
-
-            let hostname = address.split(':').next().unwrap_or(&address);
-
-            let io = tokio_openssl::connect(config, &hostname, stream).await;
-
-            debug!("Tls connection {:?} for {:?}", io, address);
-
-            io.map(|io| {
-                let stream: Pin<Box<dyn BridgeIo>> = Box::pin(io);
-                Ok((stream, password))
-            })
-            .map_err(|e| Error::new(ErrorKind::NotConnected, e))?
-        })
-    }
-}
 
 pub struct MqttClientConfig {
     addr: String,
@@ -211,7 +47,7 @@ impl MqttClientConfig {
 
 /// This is a wrapper over mqtt3 client
 pub struct MqttClient<H> {
-    client: Client<BridgeIoSource>,
+    client: Client<ClientIoSource>,
     event_handler: H,
 }
 
@@ -222,7 +58,7 @@ where
     pub fn tcp(config: MqttClientConfig, event_handler: H) -> Result<Self, ClientError> {
         let token_source = Self::token_source(&config.credentials);
         let tcp_connection = TcpConnection::new(config.addr, token_source, None);
-        let io_source = BridgeIoSource::Tcp(tcp_connection);
+        let io_source = ClientIoSource::Tcp(tcp_connection);
 
         Self::new(
             config.keep_alive,
@@ -238,7 +74,7 @@ where
 
         let token_source = Self::token_source(&config.credentials);
         let tcp_connection = TcpConnection::new(config.addr, token_source, trust_bundle);
-        let io_source = BridgeIoSource::Tls(tcp_connection);
+        let io_source = ClientIoSource::Tls(tcp_connection);
 
         Self::new(
             config.keep_alive,
@@ -254,7 +90,7 @@ where
         clean_session: bool,
         event_handler: H,
         connection_credentials: &Credentials,
-        io_source: BridgeIoSource,
+        io_source: ClientIoSource,
     ) -> Result<Self, ClientError> {
         let (client_id, username) = match connection_credentials {
             Credentials::Provider(provider_settings) => (
@@ -563,7 +399,7 @@ pub enum ClientError {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::settings::AuthenticationSettings;
+    use mqtt_util::client_io::AuthenticationSettings;
 
     #[derive(Default)]
     struct EventHandler {}
@@ -592,7 +428,7 @@ mod tests {
             clean_session,
             event_handler,
             &connection_credentials,
-            BridgeIoSource::Tcp(TcpConnection::new(addr, None, None)),
+            ClientIoSource::Tcp(TcpConnection::new(addr, None, None)),
         );
 
         assert_eq!(client.is_ok(), true);
@@ -612,7 +448,7 @@ mod tests {
             clean_session,
             event_handler,
             &connection_credentials,
-            BridgeIoSource::Tcp(TcpConnection::new(addr, None, None)),
+            ClientIoSource::Tcp(TcpConnection::new(addr, None, None)),
         );
 
         assert_eq!(client.is_err(), true);
@@ -637,7 +473,7 @@ mod tests {
             clean_session,
             event_handler,
             &connection_credentials,
-            BridgeIoSource::Tcp(TcpConnection::new(addr, None, None)),
+            ClientIoSource::Tcp(TcpConnection::new(addr, None, None)),
         );
 
         assert_eq!(client.is_err(), true);
@@ -662,7 +498,7 @@ mod tests {
             clean_session,
             event_handler,
             &connection_credentials,
-            BridgeIoSource::Tcp(TcpConnection::new(addr, None, None)),
+            ClientIoSource::Tcp(TcpConnection::new(addr, None, None)),
         );
 
         assert_eq!(client.is_err(), true);
