+++ mqtt/mqtt-bridge/src/lib.rs
@@ -19,7 +19,6 @@ mod messages;
 mod persist;
 pub mod pump;
 pub mod settings;
-mod token_source;
 pub mod upstream;
 
 pub use crate::controller::{BridgeController, BridgeControllerHandle, Error};
