+++ mqtt/mqtt-util/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "mqtt-util"
+version = "0.1.0"
+authors = ["Azure IoT Edge Devs"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1.0"
+async-trait = "0.1"
+chrono = "0.4"
+futures-util = "0.3"
+openssl = "0.10"
+percent-encoding = "1.0"
+serde = { version = "1.0", features = ["derive", "rc"] }
+thiserror = "1.0"
+tracing = "0.1"
+tokio = { version = "0.2", features = ["dns", "rt-util"] }
+tokio-openssl = "0.4"
+url = "2.1"
+
+edgelet-client = { path = "../edgelet-client" }
+mqtt3 = { path = "../mqtt3" }
